Extracts the formula used to train a pptr model.

Extracts the formula used to train a pptr model.

Usage

# S3 method for class 'pptr'
formula(x, ...)

Arguments

x

A pptr model.

...

(unused) other parameters typically passed to formula.

Value

The formula used to train the model.

See also

Examples

model <- pptr(Type ~ ., data = iris)
formula(model)
#> Type ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width - 
#>     1
#> <environment: 0x55adbbc44a60>